[USBHUB_NEW] Fix x64 warnings.

+++ drivers/usb/usbhub_new/pnp.c
@@ -1558,7 +1558,7 @@ USBH_PdoQueryId(IN PUSBHUB_PORT_PDO_EXTENSION PortExtension,
     WCHAR Buffer[200];
     PWCHAR EndBuffer;
     size_t Remaining = sizeof(Buffer);
-    ULONG Length;
+    size_t Length;
     PWCHAR Id = NULL;
     NTSTATUS Status = STATUS_SUCCESS;
     PUSB_DEVICE_DESCRIPTOR DeviceDescriptor;
@@ -1855,8 +1855,8 @@ USBH_PdoQueryDeviceText(IN PUSBHUB_PORT_PDO_EXTENSION PortExtension,
     PWCHAR DeviceText;
     UCHAR iProduct = 0;
     NTSTATUS Status;
-    ULONG NumSymbols;
-    ULONG Length;
+    size_t NumSymbols;
+    size_t Length;
 
     DPRINT("USBH_PdoQueryDeviceText ... \n");
